requirements.txt
PyQt5
pandas
Pillow
boto3
openai 